+++ et-common-libs/CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 ### Fixed
-- [SW-15924] Support simultaneous profiling from the 2 threads of the minion w/o corruptions
+- [SW-15924] Support simultaneous profiling from the 2 threads of the minion w/o corruptions (using default API entries)
 ### Security
 
 ## [0.14.0] - 2023-01-10